automation/requirements.txt
openai
moviepy
python-dotenv
pandas
requests
fal-client
apify-client
